Metrics prune script
Carbon handles retention times for each metric so the metric file keeps
a constant size. But since the metrics are created dynamically, there
may be cases where many useless metrics are created and they will be
kept forever. This causes slowness and errors in Graphite.
This script allows to detect inactive metrics and remove them. If the
metric file hasn't been modified for a number of days, it will be deleted.