Fix graphite-web slow start
The recursive chown on /var/vcap/store/graphite becomes very slow once
we collect a large number of metrics and graphite-web fails to start
because it exceeds the timeout.
It's actually not needed because all the new files are created as user
vcap by carbon. By replacing the recursive chown by individual chown on
the top directories we avoid this issue.